% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_files_to_parquet.R
\name{convert_files_to_parquet}
\alias{convert_files_to_parquet}
\title{convert_files_to_parquet}
\usage{
convert_files_to_parquet(infiles, outfile, format = "csv", delimiter = ",")
}
\arguments{
\item{infiles}{vector of file paths}

\item{outfile}{name of the outfile with an extension i.e. outfile.parquet}

\item{format}{format of the infiles}

\item{delimiter}{delimiter used in the infiles}
}
\description{
convert_files_to_parquet
}
