% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_parquet.R
\name{convert_to_parquet}
\alias{convert_to_parquet}
\title{convert_to_parquet}
\usage{
convert_to_parquet(infile, delimiter = ",")
}
\arguments{
\item{infile}{path to a file to convert}

\item{delimiter}{seperator used in the infile, defaults to csv}
}
\description{
convert_to_parquet
}
