% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_files.R
\name{glimpse_files}
\alias{glimpse_files}
\title{glimpse_files}
\usage{
glimpse_files(infiles, n = 10, delimiter = ",")
}
\arguments{
\item{infiles}{vector of file paths}

\item{n}{number of rows to read in}

\item{delimiter}{delimiter}
}
\description{
glimpse_files
}
