% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_paths.R
\name{grab_paths}
\alias{grab_paths}
\title{grab_paths}
\usage{
grab_paths(path, pattern = ".")
}
\arguments{
\item{path}{folder that contains the files to detect}

\item{pattern}{restrict files to those that match this regex pattern}
}
\description{
grab_paths
}
